Extract methods from ArrowFlightJdbcIntervalVectorAccessor class to utility class
